+++ package.json
@@ -61,8 +61,6 @@
 		"eslint-plugin-react-hooks": "^4.1.0",
 		"faker": "^4.1.0",
 		"prettier": "^2.0.5",
-		"react": "^16.13.1",
-		"react-dom": "^16.13.1",
 		"typescript": "^4.0.2"
 	}
 }
