+++ src/firestore/index.ts
@@ -1,3 +1,1 @@
-export { useWriteFS } from './write/useWriteFS';
-export * from './useReadCollection';
-export * from './read/useReadFS';
+export { useSet } from './write/useSet';
