+++ src/index.ts
@@ -1,5 +1,5 @@
 // A file for exporting APIs available to users
 export * from './context/CreateFirebaseContext';
-export * from './firestore';
 export * from './auth';
 export { createDate } from './utils/firebaseDate';
+export * from './firestore';
